Expose methods as camelCase and add attributes for controlling renaming
Add an attribute to methods to control the name they're exported under
and an attribute to php_impl to override automatic case conversion
conventions.
